+++ src/action/actionType.js
@@ -9,6 +9,3 @@ export const FILTER_RESOURCE_BY_CATEGORIES = 'FILTER_RESOURCE_BY_CATEGORIES';
 export const ADD_SAVED_RESOURCE = 'ADD_SAVED_RESOURCE';
 
 export const REMOVE_SAVED_RESOURCE = 'REMOVE_SAVED_RESOURCE';
-
-export const GROUP_RESOURCE_BY_COORDINATES_START = 'GROUP_RESOURCE_BY_COORDINATES_START';
-export const GROUP_RESOURCE_BY_COORDINATES_SUCCESS = 'GROUP_RESOURCE_BY_COORDINATES_SUCCESS';
