+++ src/components/ResultList.js
@@ -7,7 +7,6 @@ import styles from './ResultList.module.css';
 import SortBar from './Common/SortBar.js';
 import { getDistance } from '../utils/distance.js';
 import * as resourceAction from '../action/resourceDataAction';
-import * as mapResourceAction from '../action/mapDataAction';
 
 export class ResultList extends Component {
 
@@ -73,17 +72,6 @@ export class ResultList extends Component {
     }
   }
 
-  componentDidUpdate(prevProps, nextProps) {
-    if(this.props.savedResource.length){
-    this.props.mapActions.groupResourcesByCoordinates(this.props.savedResource);
-    }
-    /*var locationArray = [];
-    this.props.resource.forEach(function (resource) {
-      locationArray[resource.hashCoordinates] = locationArray[resource.hashCoordinates] || [];
-      locationArray[resource.hashCoordinates].push(resource);
-    });*/
-  }
-
   render() {
     const sortOptions = [
       {key: 'Alphabetically', sort: this.sortByAlphabet, disabled: false}
@@ -133,7 +121,6 @@ function mapStateToProps(state, ownProps) {
 function mapDispatchToProps(dispatch) {
   return {
     actions: bindActionCreators(resourceAction, dispatch),
-    mapActions: bindActionCreators(mapResourceAction, dispatch)
   };
 }
