+++ src/reducers/index.js
@@ -1,10 +1,8 @@
 import { combineReducers } from 'redux';
 import * as resources from './resourceReducer';
-import * as mapResource from './mapResourceReducer';
 
 const rootReducer = combineReducers(
-    resources,
-    mapResource
+    resources
 );
 
 export default rootReducer;
