+++ src/reducers/resourceReducer.js
@@ -55,62 +55,3 @@ export function savedResource(state = initialState.savedResource, action) {
     }
 }
 
-function groupBy(list, keyGetter) {
-    const map = new Map();
-    list.forEach((item) => {
-        const key = keyGetter(item);
-        const collection = map.get(key);
-        if (!collection) {
-            map.set(key, [item]);
-        } else {
-            collection.push(item);
-        }
-    });
-    return map;
-}
-
-export function mapResource(state = initialState.mapResource, action) {
-    switch (action.type) {
-        case types.GROUP_RESOURCE_BY_COORDINATES_SUCCESS:
-            var locationArray = [];
-            action.resources.forEach(function (resource) {
-                if (!locationArray[resource.hashCoordinates]) {
-                    locationArray[resource.hashCoordinates] = {
-                        coordinates: resource.coordinates,
-                        groupedResource: [],
-                        showInfo: false
-                    }
-                }
-                locationArray[resource.hashCoordinates].groupedResource.push(resource);
-            }
-            );
-            /*var locationArray = Object.values(action.resources.reduce((result, {
-                id,
-                coordinates,
-                combinedaddress,
-                name,
-                phone
-            }) => {
-                // Create new group
-                if (!result[coordinates]) result[coordinates] = {
-                    coordinates,
-                    isOpen: false,
-                    data : []
-                };
-                // Append to group
-                result[coordinates].data.push({
-                    id,
-                    coordinates,
-                    combinedaddress,
-                    name,
-                    phone
-                });
-                return result;
-            }, {}));*/
-            const groupedLocation = groupBy(action.resources, resource => resource.hashCoordinates);
-            console.log("Grouped locations: ", groupedLocation);
-            return Object.assign({}, state, locationArray);
-        default: return state;
-    }
-}
-
